refactor: 알림 로직 주석 추가 및 service 통합
