+++ backend/src/main/java/com/grepp/smartwatcha/app/model/notification/NotificationJpaService.java
@@ -1,4 +1,4 @@
-package com.grepp.smartwatcha.app.model.notification.service;
+package com.grepp.smartwatcha.app.model.notification;
 
 import com.grepp.smartwatcha.app.model.notification.dto.NotificationDto;
 import com.grepp.smartwatcha.app.model.notification.event.MovieStatusChangedEvent;
@@ -17,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional(transactionManager = "jpaTransactionManager")
+// 알림 관련 Service 객체
 public class NotificationJpaService {
 
     private final MovieStatusChangeJpaRepository movieStatusChangeJpaRepository;
@@ -24,6 +25,11 @@ public class NotificationJpaService {
     private final NotificationJpaRepository notificationJpaRepository;
 
     @Scheduled(cron = "0 0 6 * * *")
+    /*
+     * 알림 스케줄러
+     * 로직: 매일 오전 6시 해당 요일에 개봉하는 영화 탐색, 있다면 개봉 여부 변경 후
+     * 보고싶어요 한 유저에게 알림 발송
+     */
     public void updateReleasedMovies() {
         Boolean status = Boolean.TRUE;
 
@@ -37,6 +43,7 @@ public class NotificationJpaService {
 
     }
 
+    // 유저가 삭제하지 않은 알림만 반환하는 메서드
     public List<NotificationDto> getActiveNotificationsForUser(Long userId) {
         List<NotificationEntity> notificationEntities = notificationJpaRepository.findByUserIdAndActivated(userId, Boolean.TRUE);
         List<NotificationDto> notificationDtos = new ArrayList<>();
@@ -69,9 +76,6 @@ public class NotificationJpaService {
     }
 
     public Long countUnread(Long id) {
-
-
-
         return (Long) notificationJpaRepository.countByUserIdAndIsReadFalse(id);
     }
 }
